Added Level class to decode level files
Use third party program called "Tiled".  Create a level with width: 25 blocks and height: 19 blocks.  Use tilesheet included in assets folder to have consistent results.  

To add a player object you have  to create an objects layer and place an object where you want the player's start position to be.  In the objects layer panel you should name the object you placed "playerStart".  or else the code wont read it properly.

TODO: - Add a goal object to indicate a completion of the level